js: Add a es2015 compatible bundle
Which includes as js modules and exports them.

TODO: check whether this helps tools like rollup
to selectively bundle modules when using a statement like:

import { Carousel, Popover } from 'bootstrap';

HEADS UP!: This breaks current browserify builds that use
require('bootstrap') to load the umd-bundle.
Users that want to keep the umd bundle need to define
a "browser" statement in their package.json like:

"browser": {
  "bootstrap": "./node_modules/bootstrap/dist/js/bootstrap.js"
}

(By using the node_module as relative path it's now treated as
a local module by browserify so that it'll not use the
browserify transform nor the "main" property we specify in our package.json)